+++ src/examinations/visibility/metadata/+components/RangeSlider.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import Slider from "@material-ui/core/Slider";
+
+export const RangeSlider = ({ value, min, max, onChange }) => {
+  return (
+    <Slider
+      min={min}
+      max={max}
+      step={0.1}
+      value={[value.min, value.max]}
+      onChange={(event, newValue) =>
+        onChange({ min: newValue[0], max: newValue[1] })
+      }
+      valueLabelDisplay="auto"
+      aria-labelledby="range-slider"
+    />
+  );
+};
